requirements.txt
pyyaml
teradatasql
python-dotenv
pandas